+++ tools/tools/usbtest/usb_modem_test.c
@@ -1,4 +1,4 @@
-/* $FreeBSD$ */
+/* $FreeBSD: head/tools/tools/usbtest/usb_modem_test.c 254241 2013-08-12 09:15:33Z hselasky $ */
 /*-
  * Copyright (c) 2007-2010 Hans Petter Selasky. All rights reserved.
  *
@@ -38,9 +38,9 @@
 #include <libusb20.h>
 #include <libusb20_desc.h>
 
-#include <dev/usb/usb_endian.h>
-#include <dev/usb/usb.h>
-#include <dev/usb/usb_cdc.h>
+#include <bus/u4b/usb_endian.h>
+#include <bus/u4b/usb.h>
+#include <bus/u4b/usb_cdc.h>
 
 #include "usbtest.h"
 
@@ -215,7 +215,7 @@ usb_modem_control_ep_test(struct modem *p, uint32_t duration, uint8_t flag)
 		}
 		timersub(&sub_tv, &ref_tv, &res_tv);
 
-		if ((res_tv.tv_sec < 0) || (res_tv.tv_sec >= duration))
+		if ((res_tv.tv_sec < 0) || (res_tv.tv_sec >= (int)duration))
 			break;
 
 		LIBUSB20_INIT(LIBUSB20_CONTROL_SETUP, &setup);
@@ -333,7 +333,7 @@ usb_modem_data_stress_test(struct modem *p, uint32_t duration)
 		}
 		timersub(&sub_tv, &ref_tv, &res_tv);
 
-		if ((res_tv.tv_sec < 0) || (res_tv.tv_sec >= duration))
+		if ((res_tv.tv_sec < 0) || (res_tv.tv_sec >= (int)duration))
 			break;
 
 		libusb20_dev_process(p->usb_dev);
